refactor; remove unused module
